Remove esql-core ParsingException (#118789)
* Remove esql-core ParsingException
Move fulltext related classes: querydsl, predicate and tests
