+++ x-pack/plugin/esql/src/test/java/org/elasticsearch/xpack/esql/expression/predicate/fulltext/FullTextUtilsTests.java
@@ -4,11 +4,11 @@
  * 2.0; you may not use this file except in compliance with the Elastic License
  * 2.0.
  */
-package org.elasticsearch.xpack.esql.core.expression.predicate.fulltext;
+package org.elasticsearch.xpack.esql.expression.predicate.fulltext;
 
 import org.elasticsearch.test.ESTestCase;
-import org.elasticsearch.xpack.esql.core.ParsingException;
 import org.elasticsearch.xpack.esql.core.tree.Source;
+import org.elasticsearch.xpack.esql.parser.ParsingException;
 
 import java.util.Map;
 
@@ -28,15 +28,15 @@ public class FullTextUtilsTests extends ESTestCase {
 
     public void testColonDelimitedErrorString() {
         ParsingException e = expectThrows(ParsingException.class, () -> FullTextUtils.parseSettings("k1=v1;k2v2", source));
-        assertThat(e.getMessage(), is("line 1:3: Cannot parse entry k2v2 in options k1=v1;k2v2"));
+        assertThat(e.getMessage(), is("line 1:2: Cannot parse entry k2v2 in options k1=v1;k2v2"));
         assertThat(e.getLineNumber(), is(1));
-        assertThat(e.getColumnNumber(), is(3));
+        assertThat(e.getColumnNumber(), is(2));
     }
 
     public void testColonDelimitedErrorDuplicate() {
         ParsingException e = expectThrows(ParsingException.class, () -> FullTextUtils.parseSettings("k1=v1;k1=v2", source));
-        assertThat(e.getMessage(), is("line 1:3: Duplicate option k1=v2 detected in options k1=v1;k1=v2"));
+        assertThat(e.getMessage(), is("line 1:2: Duplicate option k1=v2 detected in options k1=v1;k1=v2"));
         assertThat(e.getLineNumber(), is(1));
-        assertThat(e.getColumnNumber(), is(3));
+        assertThat(e.getColumnNumber(), is(2));
     }
 }
